requirements.txt

streamlit
torch
torchvision
Pillow
transformers
rdflib
numpy
openai
groq
